requirements.txt
smolagents>=1.0.0
gradio>=4.0.0
huggingface-hub>=0.20.0
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
jupyter>=1.0.0
notebook>=7.0.0
markdown>=3.5.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
ddgs>=7.0.0
python-docx>=1.1.0
